Reorganized, and optimized data-grid
