+++ src/pages/ArtistDocumentsPage.jsx
@@ -1,7 +1,7 @@
 import { DropZone } from "@/components/DropZone";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
-import DataGrid from "@/components/ui/data-grid";
+import DataGrid from "@/components/data-grid/data-grid";
 import {
   Dialog,
   DialogClose,
