+++ src/pages/ProgramsPage.jsx
@@ -7,7 +7,7 @@ import {
 import { getCurrentFiscalYear, groupByIdAndField } from "@/utils/utils";
 import Spinner from "@/components/ui/Spinner";
 import { programTableColumns } from "@/utils/TableColumns";
-import DataGrid from "@/components/ui/data-grid";
+import DataGrid from "@/components/data-grid/data-grid";
 import { PROGRAMS_EDITABLE_FIELDS } from "@/constants/constants";
 
 const BUTTON_LINKS = [
